Moves render folder under engine
