+++ crates/bitang/src/engine/chart.rs
@@ -4,8 +4,8 @@ use crate::engine::camera::Camera;
 use crate::engine::compute::{Compute, Run};
 use crate::engine::draw::Draw;
 use crate::engine::generate_mip_levels::GenerateMipLevels;
-use crate::render::image::BitangImage;
-use crate::render::SIMULATION_STEP_SECONDS;
+use crate::engine::render::image::BitangImage;
+use crate::engine::render::SIMULATION_STEP_SECONDS;
 use crate::tool::{ComputePassContext, FrameContext};
 use anyhow::{bail, ensure, Result};
 use tracing::{debug, info, warn};
