+++ crates/bitang/src/engine/material.rs
@@ -1,5 +1,4 @@
-use crate::render::draw_call::DrawCall;
-use serde::Deserialize;
+use crate::engine::render::draw_call::DrawCall;
 
 pub struct Material {
     pub passes: Vec<Option<DrawCall>>,
