+++ crates/bitang/src/engine/mod.rs
@@ -8,3 +8,4 @@ pub mod render_object;
 pub mod scene;
 pub mod pass;
 pub mod material;
+pub mod render;
