+++ crates/bitang/src/engine/render/image.rs
