+++ crates/bitang/src/engine/render/mesh.rs
@@ -1,4 +1,4 @@
-use crate::render::Vertex3;
+use crate::engine::render::Vertex3;
 use crate::tool::GpuContext;
 use anyhow::Result;
 use wgpu::util::DeviceExt;
