+++ crates/bitang/src/file/object.rs
@@ -1,6 +1,6 @@
 use crate::control::{ControlId, ControlIdPartType};
 use crate::file::chart_file::ChartContext;
-use crate::{engine, file, render};
+use crate::{engine, file, engine::render};
 use anyhow::Result;
 use serde::Deserialize;
 use std::collections::HashMap;
