+++ crates/bitang/src/file/project_file.rs
@@ -1,5 +1,5 @@
 use crate::loader::resource_repository::ResourceRepository;
-use crate::{engine, render};
+use crate::{engine, engine::render};
 use crate::tool::GpuContext;
 use anyhow::Result;
 use futures::future::join_all;
