+++ crates/bitang/src/file/scene.rs
@@ -1,7 +1,7 @@
 use crate::control::{ControlId, ControlIdPartType};
 use crate::file::chart_file::ChartContext;
 use crate::file::material::Material;
-use crate::{engine, render};
+use crate::{engine, engine::render};
 use anyhow::Context;
 use serde::Deserialize;
 use std::collections::HashMap;
