+++ crates/bitang/src/loader/gltf_loader.rs
@@ -1,6 +1,6 @@
 use crate::loader::resource_repository::{SceneFile, SceneNode};
-use crate::render::mesh::Mesh;
-use crate::render::Vertex3;
+use crate::engine::render::mesh::Mesh;
+use crate::engine::render::Vertex3;
 use crate::tool::GpuContext;
 use anyhow::{Context, Result};
 use itertools::Itertools;
