+++ crates/bitang/src/loader/resource_repository.rs
@@ -8,8 +8,8 @@ use crate::loader::resource_path::ResourcePath;
 use crate::loader::shader_cache::ShaderCache;
 use crate::loader::{CHARTS_FOLDER, CHART_FILE_NAME, PROJECT_FILE_NAME};
 use crate::engine::chart::Chart;
-use crate::render::image::{BitangImage, PixelFormat};
-use crate::render::mesh::Mesh;
+use crate::engine::render::image::{BitangImage, PixelFormat};
+use crate::engine::render::mesh::Mesh;
 use crate::engine::project::Project;
 use crate::tool::GpuContext;
 use anyhow::{anyhow, ensure, Context, Result};
