+++ crates/bitang/src/main.rs
@@ -1,7 +1,6 @@
 mod control;
 mod file;
 mod loader;
-mod render;
 mod tool;
 mod engine;
